local_worker/requirements.txt
torch==2.1.2
torchaudio==2.1.2
numpy==1.26.4
pyannote.audio==3.1.1
huggingface_hub==0.25.2
requests>=2.31.0
sounddevice>=0.4.6
